Separate pages and components
Pages contains everything directly linked to in a Route. Components contains anything else that generates HTML fragments.

Split context_button into its own file.
