+++ web/src/components/audio.rs
@@ -1,8 +1,8 @@
 use crate::components::context::{AppState, UIState};
-#[cfg(not(feature = "server_build"))]
-use crate::components::downloads_tauri::start_local_file_server;
 use crate::components::gen_components::{EpisodeModal, FallbackImage};
 use crate::components::gen_funcs::format_time_rm_hour;
+#[cfg(not(feature = "server_build"))]
+use crate::pages::downloads_tauri::start_local_file_server;
 use crate::requests::episode::Episode;
 use crate::requests::pod_req::call_get_episode_id;
 use crate::requests::pod_req::FetchPodcasting2DataRequest;
