+++ web/src/components/episode_list_item.rs
@@ -1,8 +1,8 @@
 use crate::components::audio::on_play_pause;
 use crate::components::context::{AppState, ExpandedDescriptions, UIState};
-use crate::components::gen_components::{
-    on_shownotes_click, ContextButton, EpisodeModal, FallbackImage, PageType,
-};
+use crate::components::gen_components::{on_shownotes_click, EpisodeModal, FallbackImage};
+
+use crate::components::context_menu_button::{ContextButton, PageType};
 use crate::components::gen_funcs::{
     format_datetime, match_date_format, parse_date, sanitize_html_with_blank_target, use_long_press,
 };
