+++ web/src/components/host_component.rs
@@ -1,4 +1,4 @@
-use super::gen_components::LoadingModal;
+use crate::components::gen_components::LoadingModal;
 use crate::components::context::AppState;
 use crate::requests::people_req::{
     call_get_person_subscriptions, call_subscribe_to_person, call_unsubscribe_from_person,
