+++ web/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod components;
+pub mod pages;
 pub mod requests;
 
-pub use components::*;
-
+pub use crate::components::*;
