+++ web/src/pages/mod.rs
@@ -1,0 +1,27 @@
+pub(crate) mod downloads;
+#[cfg(not(feature = "server_build"))]
+pub(crate) mod downloads_tauri;
+pub(crate) mod episode;
+pub(crate) mod episode_layout;
+pub(crate) mod feed;
+pub(crate) mod history;
+pub(crate) mod home;
+#[cfg(feature = "server_build")]
+pub(crate) mod login;
+#[cfg(not(feature = "server_build"))]
+pub(crate) mod login_tauri;
+pub(crate) mod person;
+pub(crate) mod playlist_detail;
+pub(crate) mod playlists;
+pub(crate) mod podcast_layout;
+pub(crate) mod podcasts;
+pub(crate) mod queue;
+pub(crate) mod routes;
+pub(crate) mod saved;
+pub(crate) mod search;
+pub(crate) mod search_new;
+pub(crate) mod settings;
+pub(crate) mod shared_episode;
+pub(crate) mod subscribed_people;
+pub(crate) mod user_stats;
+pub(crate) mod youtube_layout;
