+++ web/src/pages/playlist_detail.rs
