+++ web/src/pages/subscribed_people.rs
@@ -1,7 +1,7 @@
-use super::app_drawer::App_drawer;
-use super::gen_components::{empty_message, Search_nav, UseScrollToTop};
+use crate::components::app_drawer::App_drawer;
 use crate::components::audio::AudioPlayer;
 use crate::components::context::{AppState, ExpandedDescriptions, UIState};
+use crate::components::gen_components::{empty_message, Search_nav, UseScrollToTop};
 use crate::components::virtual_list::VirtualList;
 use crate::requests::episode::Episode;
 use crate::requests::people_req::{self, PersonSubscription};
