+++ web/src/pages/user_stats.rs
@@ -1,14 +1,14 @@
-use super::app_drawer::App_drawer;
-use super::gen_components::Search_nav;
+use crate::components::gen_components::Search_nav;
 use crate::components::audio::AudioPlayer;
 use crate::components::context::{AppState, UIState, UserStatsStore};
 use crate::components::gen_funcs::{format_date, format_time_mins};
 use crate::requests::pod_req::call_get_pinepods_version;
 use crate::requests::stat_reqs;
+use crate::components::app_drawer::App_drawer;
+use i18nrs::yew::use_translation;
 use yew::prelude::*;
 use yew::{function_component, html, Html};
 use yewdux::prelude::*;
-use i18nrs::yew::use_translation;
 
 #[function_component(UserStats)]
 pub fn user_stats() -> Html {
@@ -127,14 +127,14 @@ pub fn user_stats() -> Html {
                                             <p class="stats-label">{i18n.t("user_stats.episodes_downloaded")}</p>
                                             <p class="stats-value">{ &stats.EpisodesDownloaded }</p>
                                         </div>
-                                        <div class={if let Some(stats) = user_stats { 
-                                            if stats.Pod_Sync_Type.as_str() == "None" { 
-                                                "stats-card col-span-1 md:col-span-3" 
-                                            } else { 
-                                                "stats-card" 
-                                            } 
-                                        } else { 
-                                            "stats-card" 
+                                        <div class={if let Some(stats) = user_stats {
+                                            if stats.Pod_Sync_Type.as_str() == "None" {
+                                                "stats-card col-span-1 md:col-span-3"
+                                            } else {
+                                                "stats-card"
+                                            }
+                                        } else {
+                                            "stats-card"
                                         }}>
                                             <p class="stats-label">{i18n.t("user_stats.podcast_sync_status")}</p>
                                             {
